Large Eclipse-3.3 update.
git-svn-id: https://overlays.gentoo.org/svn/proj/java/java-experimental@5470 fe387f86-7aec-0310-8f44-a148a97195aa

+++ dev-util/eclipse-sdk/eclipse-sdk-3.3-r2.ebuild
@@ -20,7 +20,7 @@ SRC_URI="http://download.eclipse.org/eclipse/downloads/drops/${DMF}/${MY_A}"
 
 SLOT="3.3"
 LICENSE="EPL-1.0"
-IUSE="branding tomcat"
+IUSE="doc tomcat"
 KEYWORDS="~amd64 ~x86"
 
 S=${WORKDIR}
@@ -30,26 +30,29 @@ ECLIPSE_DIR="/usr/lib/eclipse-${SLOT}"
 
 CDEPEND=">=dev-java/ant-eclipse-ecj-3.3
 	>=dev-java/ant-core-1.7.0-r1
+	>=dev-java/swt-${SLOT}-r1
+	=dev-java/junit-3*
+	=dev-java/junit-4*
+	>=dev-java/jsch-0.1.34-r1
+	>=dev-java/icu4j-3.6.1
 	tomcat? (
 		dev-java/commons-digester-rss
 		=dev-java/mx4j-core-3*
 		=dev-java/jakarta-regexp-1.4*
 		=dev-java/servletapi-2.4*
 		>=www-servers/tomcat-5.5.17
-	)
-	>=dev-java/swt-${SLOT}-r1
-	=dev-java/junit-3*
-	=dev-java/junit-4*
-	>=dev-java/jsch-0.1.34-r1"
+	)"
 	
 RDEPEND=">=virtual/jre-1.5
 	${CDEPEND}"
+	
 DEPEND="=virtual/jdk-1.5*
 	dev-java/ant-nodeps
-	${CDEPEND}
 	>=sys-apps/findutils-4.1.7
 	app-arch/unzip
-	app-arch/zip"
+	app-arch/zip
+	${CDEPEND}"
+		
 #	=dev-java/lucene-2.2*
 #	=dev-java/lucene-analyzers-2.2*
 
@@ -62,26 +65,17 @@ pkg_setup() {
 	check_reqs
 
 	eclipse_arch=${ARCH}
-	use amd64 && eclipse_arch=x86_64
+	use amd64 && eclipse_arch="x86_64"
+	javadoc="false"
+	use doc && javadoc="true"
 }
 
 src_unpack() {
 	unpack ${A}
-
 	patch-apply
+	remove-bundled-stuff
 
-	# Remove pre-built eclipse binaries
-	find ${S} -type f -name eclipse | xargs rm
-	# ...  .so libraries
-	find ${S} -type f -name '*.so' | xargs rm
-	# ... .jar files
-	rm plugins/org.eclipse.swt/extra_jars/exceptions.jar \
-		plugins/org.eclipse.osgi/osgi/osgi*.jar \
-		plugins/org.eclipse.osgi/supplement/osgi/osgi.jar
-		
-	rm -rf plugins/org.eclipse.swt.*
-
-	# no warnings / java5 / all output should be directed to stdout
+	# No warnings / java5 / all output should be directed to stdout
 	find ${S} -type f -name '*.xml' -exec \
 		sed -r -e "s:(-encoding ISO-8859-1):\1 -nowarn:g" \
 			-e "s:(\"compilerArg\" value=\"):\1-nowarn :g" \
@@ -93,18 +87,12 @@ src_unpack() {
 	sed -r -e "s:^(JAVA_HOME =) .*:\1 $(java-config --jdk-home):" \
 		-e "s:gcc :gcc ${CFLAGS} :" \
 		-i plugins/org.eclipse.core.filesystem/natives/unix/linux/Makefile \
-		|| die "sed makefile failed"
+		|| die "sed Makefile failed"
 
-	ebegin "Symlinking system jars"
-		symlink-ant
-		symlink-junit
-		symlink-jsch
-#		symlink-lucene		
-		use tomcat && symlink-tomcat
-	eend $?
-	
+	install-link-system-jars
+
 	while read line; do
-		java-ant_rewrite-classpath "$line"
+		java-ant_rewrite-classpath "$line" > /dev/null
     done < <(find ${S} -type f -name 'build.xml' )
 }
 
@@ -114,52 +102,60 @@ src_compile() {
 	einfo "Using bootclasspath ${bootclasspath}"
 
 	java-pkg_force-compiler ecj-3.3
-	ANT_OPTS=-Xmx1024M \
-		eant -q -Dnobootstrap=true -Dlibsconfig=true -Dbootclasspath=${bootclasspath} \
-		-DinstallOs=linux -DinstallWs=gtk -DinstallArch=${eclipse_arch} \
-		-Djava5.home=$(java-config --jdk-home) -Dgentoo.classpath=`java-pkg_getjars swt-3`
+	
+	local system_jars="$(java-pkg_getjars swt-3,icu4j-3.6,ant-core,jsch):$(java-pkg_getjars --build-only ant-nodeps)"
+	local gentoo_jars="$(java-pkg_getjars ant-core,icu4j-3.6,jsch)"
+		
+	ANT_OPTS=-Xmx1024M ANT_TASKS="ant-nodeps" eant -q -Dnobootstrap=true -Dlibsconfig=true -Dbootclasspath=${bootclasspath} \
+		-DinstallOs=linux -DinstallWs=gtk -DinstallArch=${eclipse_arch} -Djavadoc="${javadoc}"\
+		-Djava5.home=$(java-config --jdk-home) -Dgentoo.classpath="${system_jars}" -Dgentoo.jars="${gentoo_jars//:/,}"
 }
 
 src_install() {
 	dodir /usr/lib
 
-	[[ -f result/linux-gtk-${eclipse_arch}-sdk.tar.gz ]] \
-		|| die "tar.gz bundle was not built properly!"
-	tar zxf result/linux-gtk-${eclipse_arch}-sdk.tar.gz -C ${D}/usr/lib \
-		|| die "Failed to extract the built package"
+	[[ -f result/linux-gtk-${eclipse_arch}-sdk.tar.gz ]] || die "tar.gz bundle was not built properly!"
+	tar zxf "result/linux-gtk-${eclipse_arch}-sdk.tar.gz" -C "${D}/usr/lib" || die "Failed to extract the built package"
 
-	mv ${D}/usr/lib/eclipse ${D}/${ECLIPSE_DIR}
+	mv "${D}/usr/lib/eclipse" "${D}/${ECLIPSE_DIR}"
 
-	# What does this do?
+	# The following moves the loader and built shared libraries to the correct place
 	
-	mv ${S}/launchertmp/eclipse ${D}/${ECLIPSE_DIR}
+	mv "${S}/launchertmp/eclipse" "${D}/${ECLIPSE_DIR}"
 	mv ${S}/launchertmp/library/gtk/eclipse_*.so ${D}/${ECLIPSE_DIR}/plugins/org.eclipse.equinox.launcher.gtk.linux.*/
 
 	# Install startup script
 	exeinto /usr/bin
-	doexe ${FILESDIR}/eclipse-${SLOT}
-	echo "-Xss2048k" >> ${D}/${ECLIPSE_DIR}/eclipse.ini
+	doexe "${FILESDIR}/eclipse-${SLOT}"
+	
+	cat > "${D}/${ECLIPSE_DIR}/eclipse.ini" <<-EOF
+-showsplash
+org.eclipse.platform
+-vmargs
+-Djava.library.path=/usr/lib
+-Xms128m
+-Xmx256m
+-Xss2048k
+EOF
 
 	make_desktop_entry eclipse-${SLOT} "Eclipse ${PV}" "${ECLIPSE_DIR}/icon.xpm"
 
-	cd ${D}/${ECLIPSE_DIR}
+	cd "${D}/${ECLIPSE_DIR}"
 	install-link-system-jars
 }
 
 pkg_postinst() {
 	einfo
-	einfo "Users can now install plugins via Update Manager without any"
-	einfo "tweaking."
+	einfo "Welcome to Eclipse-3.3 (Europa)!"
 	einfo
-	einfo "Eclipse plugin packages (ie eclipse-cdt) will likely go away in"
-	einfo "the near future until they can be properly packaged. Update Manager"
-	einfo "is prefered in the meantime."
+	einfo "You can now install plugins via Update Manager without any"
+	einfo "tweaking. This is the recommended way to install new features for Eclipse."
 	einfo
-	einfo "If you plan to use heavy plugins, mind to read"
+	einfo "If your Eclipse install seems slow, and is unstable,"
+	einfo "read (especially if you are on amd64 or use a lot of plugins):"
 	einfo "file://${ECLIPSE_DIR}/readme/readme_eclipse.html#Running%20Eclipse"
-	einfo "about memory issues"
-	einfo "(and to eventually modify ${ECLIPSE_DIR}/eclipse.ini)"
-	einfo
+	einfo "about memory issues (modify ${ECLIPSE_DIR}/eclipse.ini)"
+	
 
 	if use gcj ; then
 		rm -f /usr/lib/eclipse-${ECLIPSE_VER}/eclipse.gcjdb
@@ -172,27 +168,16 @@ pkg_postinst() {
 # -----------------------------------------------------------------------------
 
 install-link-system-jars() {
-	#local ant_dir="$(basename plugins/org.apache.ant_*)"
-	#dosym /usr/share/ant-core ${ECLIPSE_DIR}/plugins/${ant_dir}
 
 	pushd plugins/	
-	rm -rf org.apache.ant_*
 	java-pkg_jarfrom swt-3
-	java-pkg_jarfrom ant-core	
-	popd
-
-#	symlink-lucene
-	symlink-junit
-	symlink-jsch
-	use tomcat && symlink-tomcat
-}
-
-symlink-ant() {
-	pushd plugins/org.apache.ant_*/lib/
-	rm *.jar
 	java-pkg_jarfrom ant-core
-	java-pkg_jarfrom --build-only ant-nodeps
+	java-pkg_jarfrom icu4j-3.6
+	java-pkg_jarfrom jsch
 	popd
+
+	symlink-junit
+	use tomcat && symlink-tomcat
 }
 
 symlink-tomcat() {
@@ -244,18 +229,10 @@ symlink-junit() {
 	popd
 }
 
-symlink-jsch() {
-	pushd plugins/
-	local jsch_jar="$(basename com.jcraft.jsch_*.jar)"
-	rm ${jsch_jar}
-	java-pkg_jar-from jsch jsch.jar ${jsch_jar}
-	popd
-}
-
 patch-apply() {
-	# patch launcher source
+	# Patch launcher source
 	mkdir launchertmp
-	unzip -qq -d launchertmp plugins/org.eclipse.platform/launchersrc.zip >/dev/null || die "unzip failed"
+	unzip -qq -d launchertmp plugins/org.eclipse.platform/launchersrc.zip > /dev/null || die "unzip failed"
 	pushd launchertmp/
 	epatch ${PATCHDIR}/launcher_double-free.diff
 	sed -i "s/CFLAGS\ =\ -O\ -s\ -Wall/CFLAGS = ${CFLAGS}\ -Wall/" \
@@ -266,16 +243,16 @@ patch-apply() {
 	mv launchersrc.zip plugins/org.eclipse.platform/launchersrc.zip
 	rm -rf launchertmp
 
-	# disable swt, jdk6
-	epatch ${PATCHDIR}/eclipse_disable-swt.diff
-	epatch ${PATCHDIR}/eclipse_disable-jdt-tool.diff
-	epatch ${PATCHDIR}/eclipse_disable-jdk6.diff
+	# Disable swt, jdk6
+	epatch "${PATCHDIR}/eclipse_disable-swt.diff"
+	epatch "${PATCHDIR}/eclipse_disable-jdt-tool.diff"
+	epatch "${PATCHDIR}/eclipse_disable-jdk6.diff"
 
 	# What does this do?
-	epatch ${PATCHDIR}/how_to_loose_sanity_on_freaky_env_vars_argh.diff
+	epatch "${PATCHDIR}/eclipse-${SLOT}-main-build.xml.diff"
 
 	# JNI
-	epatch ${FEDORA}/eclipse-libupdatebuild2.patch
+	epatch "${FEDORA}/eclipse-libupdatebuild2.patch"
 
 	# Fedora does not apply this anymore because they checkout
 	# org.eclipse.equinox.initializer project from cvs. 'till a fix, we'll
@@ -299,7 +276,7 @@ patch-apply() {
 			-i assemble.*.xml
 	fi
 
-	# generic releng plugins that can be used to build plugins
+	# Generic releng plugins that can be used to build plugins
 	# https://www.redhat.com/archives/fedora-devel-java-list/2006-April/msg00048.html
 	pushd plugins/org.eclipse.pde.build >/dev/null
 	# %patch53
@@ -308,11 +285,28 @@ patch-apply() {
 		-i templates/package-build/build.properties
 	popd >/dev/null
 
-	# BRANDING
-# FIXME link?
-#	if use branding; then
-#		pushd plugins/org.eclipse.platform >/dev/null
-#		cp ${WORKDIR}/splash.bmp .
-#		popd >/dev/null
-#	fi
+	sed -i "s/<copy.*com\.jcraft\.jsch.*\/>//" "package.org.eclipse.sdk.linux.gtk.${eclipse_arch}.xml"
+	sed -i "s/<copy.*com\.ibm\.icu.*\/>//" "package.org.eclipse.sdk.linux.gtk.${eclipse_arch}.xml"
+	
+	#epatch "${FILESDIR}/${SLOT}/${P}-buildJSPs.patch"
+	#cp "${FILESDIR}/${SLOT}/buildJSPs.xml" "plugins/org.eclipse.help.webapp/buildJSPs.xml"
+	cp "${FILESDIR}/${SLOT}/isv-build.xml" "plugins/org.eclipse.help.webapp/build.xml"
+	#cp "${FILESDIR}/${SLOT}/build-isv.xml" "plugins/org.eclipse.platform.doc.isv/build.xml"
+	cp "${FILESDIR}/${SLOT}/buildDoc.xml" "plugins/org.eclipse.platform.doc.isv/buildDoc.xml"	
+}
+
+remove-bundled-stuff() {
+	# Remove pre-built eclipse binaries
+	find ${S} -type f -name eclipse | xargs rm
+	# ...  .so libraries
+	find ${S} -type f -name '*.so' | xargs rm
+	# ... .jar files
+	rm plugins/org.eclipse.swt/extra_jars/exceptions.jar \
+		plugins/org.eclipse.osgi/osgi/osgi*.jar \
+		plugins/org.eclipse.osgi/supplement/osgi/osgi.jar
+		
+	rm -rf plugins/org.eclipse.swt.*
+	rm -rf plugins/org.apache.ant_*
+	rm plugins/com.jcraft.jsch*
+	rm plugins/com.ibm.icu*
 }
